requirements.txt
tensorflow==2.2.0
transformers==3.0.2
sentence-transformers==0.3.4
sentencepiece==0.1.91
scipy==1.4.1
scikit-learn==0.24.1
flair==0.6.1.post1
nlp==0.3.0
pyarrow==0.17.1

wandb
pandas
bert-score
visdom
tensorboardX
tensorflow_hub
nltk
numpy
pymemcache
mezmorize
cached_property
requests
lru-dict
python-Levenshtein

# TextAttack
lemminflect
language_tool_python
editdistance
filelock
terminaltables
tqdm
word2number
num2words

# xu_auto_LiRPA
appdirs>=1.4
oslo.concurrency>=4.2
pytest>=5.0
pytorch_pretrained_bert
